refactor(onboarding): extract ViewDemoProjectButton to separate file
